Rename package name to prepare
